+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Added MBB Platform (used in e.g. e-up)
 - Duplicate log entry filter
+- Added vehicle warning lights
 
 ### Changed
 - Changed to new selective update endpoint. Warning this is a breaking change in the API!